requirements.txt
python-telegram-bot[job-queue]==21.6

httpx==0.27.2
requests==2.31.0

beautifulsoup4==4.12.3
lxml==5.2.1

apscheduler
pytz